+++ protobuf/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.13.6
+
+* Override `operator ==` and `hashCode` in `PbMap` so that two `PbMap`s are equal if they have equal key/value pairs.
+
 ## 0.13.5
 
 * Add new method `addAll` on ExtensionRegistry for more conveniently adding multiple extensions at once.